+++ test/writer_SUITE.erl
@@ -30,13 +30,17 @@
   send_events_pid/1,
   send_events_tcp/1,
   send_events_udp/1,
+  send_events_async/1,
   send_state/1,
   send_states/1,
   send_states_pid/1,
   send_states_tcp/1,
   send_states_udp/1,
+  send_states_async/1,
   send_entities/1,
-  send_entities_pid/1
+  send_entities_pid/1,
+  send_entities_async/1,
+  ignore_unknown_messages/1
 ]).
 
 % Common Test
@@ -48,22 +52,26 @@ all() ->
     send_events_pid,
     send_events_tcp,
     send_events_udp,
+    send_events_async,
     send_state,
     send_states,
     send_states_pid,
     send_states_tcp,
     send_states_udp,
+    send_states_async,
     send_entities,
-    send_entities_pid
+    send_entities_pid,
+    send_entities_async,
+    ignore_unknown_messages
   ].
 
 init_per_suite(Config) ->
   ok = katja:start(),
   Config.
 
-init_per_testcase(Test, Config) when Test =:= send_events_pid; Test =:= send_events_tcp; Test =:= send_events_udp;
-                                     Test =:= send_states_pid; Test =:= send_states_tcp; Test =:= send_states_udp;
-                                     Test =:= send_entities_pid ->
+init_per_testcase(Test, Config) when Test =:= send_events_pid; Test =:= send_events_tcp; Test =:= send_events_udp; Test =:= send_events_async;
+                                     Test =:= send_states_pid; Test =:= send_states_tcp; Test =:= send_states_udp; Test =:= send_states_async;
+                                     Test =:= send_entities_pid; Test =:= send_entities_async; Test =:= ignore_unknown_messages ->
   {ok, WPid} = katja_writer:start_link(),
   [{pid_writer, WPid} | Config];
 init_per_testcase(_Test, Config) ->
@@ -73,9 +81,9 @@ end_per_suite(_Config) ->
   ok = katja:stop(),
   ok.
 
-end_per_testcase(Test, Config) when Test =:= send_events_pid; Test =:= send_events_tcp; Test =:= send_events_udp;
-                                    Test =:= send_states_pid; Test =:= send_states_tcp; Test =:= send_states_udp;
-                                    Test =:= send_entities_pid ->
+end_per_testcase(Test, Config) when Test =:= send_events_pid; Test =:= send_events_tcp; Test =:= send_events_udp; Test =:= send_events_async;
+                                    Test =:= send_states_pid; Test =:= send_states_tcp; Test =:= send_states_udp; Test =:= send_states_async;
+                                    Test =:= send_entities_pid; Test =:= send_entities_async; Test =:= ignore_unknown_messages ->
   WPid = ?config(pid_writer, Config),
   ok = katja_writer:stop(WPid),
   ok;
@@ -121,6 +129,19 @@ send_events_udp(Config) ->
   ok = katja:send_events(katja_writer, udp, [Event, Event]),
   ok = katja:send_events(WPid, udp, [Event, Event]).
 
+send_events_async(Config) ->
+  WPid = ?config(pid_writer, Config),
+  SmallEvent = [{service, "katja 1"}, {metric, 9001}],
+  ok = katja:send_event_async(SmallEvent),
+  ok = katja:send_event_async(WPid, tcp, SmallEvent),
+  ok = katja:send_event_async(WPid, SmallEvent),
+  Description = lists:flatten(lists:duplicate(4096, "abcd")),
+  Event = [{service, "katja 1"}, {metric, 9001}, {description, Description}],
+  ok = katja:send_event_async(WPid, Event),
+  ok = katja:send_events_async([Event, Event]),
+  ok = katja:send_events_async(WPid, [Event, Event]),
+  ok = katja:send_events_async(WPid, tcp, [Event, Event]).
+
 send_state(_Config) ->
   ok = katja:send_state([{service, "katja 1"}, {state, "testing"}]),
   ok = katja:send_state([{service, "katja 1"}, {state, "testing"}, {ttl, 60}]).
@@ -149,6 +170,17 @@ send_states_udp(Config) ->
   ok = katja:send_states(katja_writer, udp, [State, State]),
   ok = katja:send_states(WPid, udp, [State, State]).
 
+send_states_async(Config) ->
+  WPid = ?config(pid_writer, Config),
+  State = [{service, "katja 1"}, {state, "testing"}],
+  ok = katja:send_state_async(State),
+  ok = katja:send_state_async(WPid, tcp, State),
+  ok = katja:send_state_async(WPid, State),
+  ok = katja:send_states_async([State, State]),
+  ok = katja:send_states_async(katja_writer, [State, State]),
+  ok = katja:send_states_async(WPid, [State, State]),
+  ok = katja:send_states_async(WPid, udp, [State, State]).
+
 send_entities(_Config) ->
   Description = lists:flatten(lists:duplicate(4096, "abcd")),
   Event = [{service, "katja 1"}, {metric, 9001}, {description, Description}],
@@ -164,3 +196,18 @@ send_entities_pid(Config) ->
   ok = katja:send_entities(WPid, [{events, [Event]}]),
   ok = katja:send_entities(WPid, [{states, [State]}]),
   ok = katja:send_entities(WPid, [{states, [State, State]}, {events, [Event, Event]}]).
+
+send_entities_async(Config) ->
+  WPid = ?config(pid_writer, Config),
+  Event = [{service, "katja 1"}, {metric, 9001}],
+  State = [{service, "katja 1"}, {state, "testing"}],
+  ok = katja:send_entities_async(WPid, [{events, [Event]}]),
+  ok = katja:send_entities_async(WPid, [{states, [State]}]),
+  ok = katja:send_entities_async([{states, [State, State]}, {events, [Event, Event]}]),
+  ok = katja:send_entities_async(WPid, [{states, [State, State]}, {events, [Event, Event]}]).
+
+ignore_unknown_messages(Config) ->
+  WPid = ?config(pid_writer, Config),
+  ignored = gen_server:call(WPid, foobar),
+  ok = gen_server:cast(WPid, foobar),
+  WPid ! foobar.